fix: resolve sign-conversion and int-float-conversion warnings, fix .clang-tidy syntax
- PerformanceFile.cpp: add static_cast<size_t> for MemoryBlock ctor
- KeyboardPanel.cpp: add static_cast<float> before int×float multiply
- .clang-tidy: fix modernize-use-using.IgnoreExternC YAML syntax
- Apply clang-format on affected files

+++ source/Recording/PerformanceFile.cpp
@@ -30,7 +30,7 @@ RecordingEventSource stringToSource(const juce::String& str) {
 // --- MidiMessage <-> var ---
 
 juce::var midiMessageToVar(const juce::MidiMessage& msg) {
-    juce::MemoryBlock mb(msg.getRawData(), msg.getRawDataSize());
+    juce::MemoryBlock mb(msg.getRawData(), static_cast<size_t>(msg.getRawDataSize()));
     return juce::var(mb.toBase64Encoding());
 }
